Moved Mock Queue.Item out into test/java
Not sure why I ever had it in Queue.MockItem. I must have wrongly though it was private/protected in some way that did not allow it to live anywhere else.
